build: breaking down stripe service
